gbm: Use unsigned for BO offset getter
The actual offset returned is uint32_t, however int64_t was used as the
return type from gbm_bo_get_offset to allow negative returns to signal
errors to the caller.

In case of an error getting the offset, the user will also be unable to
get the handle/FD, and thus have nothing to offset into. This means that
returning 0 as an error value is harmless, allowing us to change the
return type to uint32_t in order to avoid signed/unsigned confusion in
callers.

Cc: Ben Widawsky <ben@bwidawsk.net>
Cc: Jason Ekstrand <jason@jlekstrand.net>

+++ src/gbm/main/gbm.c
@@ -203,7 +203,7 @@ gbm_bo_get_format(struct gbm_bo *bo)
  * \param bo The buffer object
  * \return The offset
  */
-GBM_EXPORT int64_t
+GBM_EXPORT uint32_t
 gbm_bo_get_offset(struct gbm_bo *bo, int plane)
 {
    return bo->gbm->bo_get_offset(bo, plane);
